requirements.txt
Django==4.2.4
django-tinymce==3.6.1
django-phonenumber-field==7.1.0
Pillow==10.0.0
